Added support for blinking
Added env var EDMTMPFILES - location of autosave files (default is still /tmp)

Private color maps may now be used with version 3 & 4 color files (might
be useful for 8 bit displays)

+++ lib/entry_form.cc
@@ -2790,7 +2790,7 @@ long value;
 //   printf( "old value = %-d\n", destArray[i] );
 //   printf( "new value = %-d\n", value );
 
-  destArray[i] = value;
+  destArray[i] = (int) value;
 
 }
